Translate all quiz-related templates to Spanish for both students and faculty. Standardize visible UI text, improve clarity, and fix minor inconsistencies in labels and instructions. Update all quiz pages, including creation, attempt, results, summary, and question management, to ensure a fully localized experience.

+++ MetaMath/quiz/models.py
@@ -52,7 +52,7 @@ class Quiz(models.Model):
 class Question(models.Model):
     quiz = models.ForeignKey(Quiz, on_delete=models.CASCADE)
     question = models.TextField()
-    marks = models.IntegerField(default=0, null=False)
+    marks = models.DecimalField(max_digits=5, decimal_places=2, default=0, null=False)  # Cambiado aquí
     option1 = models.TextField(null=False, blank=False, default='',)
     option2 = models.TextField(null=False, blank=False, default='')
     option3 = models.TextField(null=False, blank=False, default='')
